Move cove-api from apps/ to services/
Backend services now live under services/ (reserved for this from the
start), keeping apps/ for client applications (iOS, future web). cove-api
is the first service to make the move while it's the only one — the
cheapest moment to settle the convention.

Updates the Go module path and internal imports, root Makefile targets,
both CI workflows (path filters, working dirs, docker context), and the
iOS openapi.yaml sync-comment. The Artifact Registry path was already
'services/cove-api' so published image tags are unchanged. No homelab
manifest, database, or runtime config changes.
