+++ services/cove-api/.gitignore
