+++ services/cove-api/internal/auth/middleware.go
